requirements.txt
# ==========================================================================
# InvIQ Smart Inventory Assistant - Python Dependencies
# ==========================================================================

# -- Web Framework ---------------------------------------------------------
fastapi==0.104.1
uvicorn[standard]==0.24.0
gunicorn==21.2.0
python-multipart==0.0.6

# -- Database --------------------------------------------------------------
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1

# -- Security & Auth -------------------------------------------------------
pyjwt==2.8.0
pwdlib[argon2]==0.2.0
email-validator==2.1.0

# -- Caching & Rate Limiting -----------------------------------------------
redis==5.0.1
upstash-redis==1.1.0
slowapi==0.1.9

# -- AI & LLM -------------------------------------------------------------
langchain>=1.2.15
langchain-groq>=1.1.2
langchain-community>=0.4.1
langchain-openai>=1.2.1
langchain-core>=1.3.2
langsmith>=0.7.37
langgraph>=1.1.10
chromadb==0.4.22
sentence-transformers==2.2.2

# -- Data Processing -------------------------------------------------------
pydantic==2.5.2
pydantic-settings==2.1.0
openpyxl==3.1.2

# -- PDF Generation --------------------------------------------------------
reportlab==4.0.7

# -- HTTP & Observability --------------------------------------------------
httpx==0.25.2
requests==2.31.0

# -- WebSocket -------------------------------------------------------------
websockets==12.0

# -- Utilities -------------------------------------------------------------
python-dotenv==1.0.0

# -- Testing (Development Only) --------------------------------------------
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0

# -- Code Quality (Development Only) ---------------------------------------
ruff==0.1.8
black==23.12.1
mypy==1.7.1